move implementation to cpp
