+++ cloud/storage/core/libs/netlink/message.h
@@ -14,19 +14,8 @@ private:
     nlattr* Attribute;
 
 public:
-    TNestedAttribute(nl_msg* message, int attribute)
-        : Message(message)
-    {
-        Attribute = nla_nest_start(message, attribute);
-        if (!Attribute) {
-            throw TServiceError(E_FAIL) << "unable to nest attribute";
-        }
-    }
-
-    ~TNestedAttribute()
-    {
-        nla_nest_end(Message, Attribute);
-    }
+    TNestedAttribute(nl_msg* message, int attribute);
+    ~TNestedAttribute();
 };
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -37,32 +26,15 @@ private:
     nl_msg* Message;
 
 public:
-    TMessage(int family, int command, int flags = 0, int version = 0)
-    {
-        Message = nlmsg_alloc();
-        if (Message == nullptr) {
-            throw TServiceError(E_FAIL) << "unable to allocate message";
-        }
-        genlmsg_put(
-            Message,
-            NL_AUTO_PORT,
-            NL_AUTO_SEQ,
-            family,
-            0,   // hdrlen
-            flags,
-            command,
-            version);
-    }
-
-    ~TMessage()
-    {
-        nlmsg_free(Message);
-    }
+    TMessage(int family, int command, int flags = 0, int version = 0);
+    ~TMessage();
 
     operator nl_msg*() const
     {
         return Message;
     }
+
+    TNestedAttribute Nest(int attribute);
 
     template <typename T>
     void Put(int attribute, T data)
@@ -72,11 +44,6 @@ public:
                 << "unable to put attribute " << attribute << ": "
                 << nl_geterror(err);
         }
-    }
-
-    TNestedAttribute Nest(int attribute)
-    {
-        return TNestedAttribute(Message, attribute);
     }
 };
