+++ cloud/storage/core/libs/netlink/ya.make
@@ -5,6 +5,11 @@ LICENSE_RESTRICTION_EXCEPTIONS(
     contrib/restricted/libnl/lib/nl-genl-3
 )
 
+SRCS(
+    message.cpp
+    socket.cpp
+)
+
 PEERDIR(
     cloud/storage/core/libs/common
     contrib/restricted/libnl/lib/nl-3
